PrevailingTrend/第一层模块/国外热点数据/requirements.txt
# 数据库连接
pymysql>=1.0.0
sqlalchemy>=1.4.0

# 数据处理
pandas>=1.3.0
numpy>=1.21.0

# HTTP请求
requests>=2.25.0
aiohttp>=3.8.0

# Web框架
flask>=2.0.0
flask-cors>=3.0.10

# 配置管理
python-dotenv>=0.19.0

# 日志管理
loguru>=0.5.0

# 网络爬虫
beautifulsoup4>=4.9.0
lxml>=4.6.0

# 文本处理
jieba>=0.42.1 